# optimize CSG.union for array input